from class 2/10

+++ APCS-A_2020/src/unit2/Line.java
@@ -7,6 +7,6 @@ package unit2;
 
 public class Line {
 	public static double getSlope(int x1, int y1, int x2, int y2) {
-		return (y2 - y1) / (x2 - x1);
+		 return (double) (y2 - y1) / (x2 - x1);
 	}
 }
